requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-HTTPAuth==4.8.0
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
playwright==1.58.0
pyee==13.0.1
typing_extensions==4.15.0
Werkzeug==3.1.6